Refactor postgres query wizard
