+++ src/postgres/commands/PostgresQueryWizards/FunctionQueryWizard/FunctionQueryNameStep.ts
@@ -4,8 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AzureWizardPromptStep } from "vscode-azureextensionui";
-import { ext } from "../../../extensionVariables";
-import { localize } from "../../../utils/localize";
+import { ext } from "../../../../extensionVariables";
+import { localize } from "../../../../utils/localize";
+import { validateIdentifier } from "../validateIdentifier";
 import { IPostgresFunctionQueryWizardContext } from "./IPostgresFunctionQueryWizardContext";
 
 export class FunctionQueryNameStep extends AzureWizardPromptStep<IPostgresFunctionQueryWizardContext> {
@@ -19,23 +20,4 @@ export class FunctionQueryNameStep extends AzureWizardPromptStep<IPostgresFuncti
     public shouldPrompt(wizardContext: IPostgresFunctionQueryWizardContext): boolean {
         return !wizardContext.name;
     }
-}
-
-function validateIdentifier(identifier: string): string | undefined {
-    // Identifier naming rules: https://aka.ms/AA8618j
-    identifier = identifier.trim();
-
-    if (!identifier) {
-        return localize('cannotBeEmpty', 'Name cannot be empty.');
-    }
-
-    if (!identifier[0].match(/[a-z_]/i)) {
-        return localize('mustStartWithLetterOrUnderscore', 'Name must start with a letter or underscore.');
-    }
-
-    if (identifier.match(/[^a-z_\d$]/i)) {
-        return localize('canOnlyContainCertainCharacters', 'Name can only contain letters, underscores, digits (0-9), and dollar signs ($).');
-    }
-
-    return undefined;
 }
