+++ src/postgres/commands/PostgresQueryWizards/FunctionQueryWizard/FunctionQueryReturnTypeStep.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AzureWizardPromptStep, IAzureQuickPickItem, IWizardOptions } from "vscode-azureextensionui";
-import { ext } from "../../../extensionVariables";
-import { localize } from "../../../utils/localize";
+import { ext } from "../../../../extensionVariables";
+import { localize } from "../../../../utils/localize";
 import { FunctionQueryCustomReturnTypeStep } from "./FunctionQueryCustomReturnTypeStep";
 import { IPostgresFunctionQueryWizardContext } from "./IPostgresFunctionQueryWizardContext";
