+++ src/postgres/commands/PostgresQueryWizards/FunctionQueryWizard/IPostgresFunctionQueryWizardContext.ts
@@ -3,10 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IActionContext } from "vscode-azureextensionui";
+import { IPostgresQueryWizardContext } from "../IPostgresQueryWizardContext";
 
-export interface IPostgresFunctionQueryWizardContext extends IActionContext {
-    name?: string;
+export interface IPostgresFunctionQueryWizardContext extends IPostgresQueryWizardContext {
     returnType?: string;
-    query?: string;
 }
