+++ src/postgres/commands/PostgresQueryWizards/IPostgresQueryWizardContext.ts
@@ -1,0 +1,12 @@
+
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IActionContext } from "vscode-azureextensionui";
+
+export interface IPostgresQueryWizardContext extends IActionContext {
+    name?: string;
+    query?: string;
+}
