+++ src/postgres/commands/createPostgresFunctionQuery.ts
@@ -4,15 +4,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AzureWizard, IActionContext } from "vscode-azureextensionui";
-import { postgresBaseFileName, postgresFileExtension } from "../../constants";
-import { nonNullProp } from "../../utils/nonNull";
-import * as vscodeUtil from '../../utils/vscodeUtils';
 import { PostgresFunctionsTreeItem } from "../tree/PostgresFunctionsTreeItem";
-import { connectPostgresDatabase } from "./connectPostgresDatabase";
-import { FunctionQueryCreateStep } from "./FunctionQueryWizard/FunctionQueryCreateStep";
-import { FunctionQueryNameStep } from "./FunctionQueryWizard/FunctionQueryNameStep";
-import { FunctionQueryReturnTypeStep } from "./FunctionQueryWizard/FunctionQueryReturnTypeStep";
-import { IPostgresFunctionQueryWizardContext } from "./FunctionQueryWizard/IPostgresFunctionQueryWizardContext";
+import { FunctionQueryCreateStep } from "./PostgresQueryWizards/FunctionQueryWizard/FunctionQueryCreateStep";
+import { FunctionQueryNameStep } from "./PostgresQueryWizards/FunctionQueryWizard/FunctionQueryNameStep";
+import { FunctionQueryReturnTypeStep } from "./PostgresQueryWizards/FunctionQueryWizard/FunctionQueryReturnTypeStep";
+import { IPostgresFunctionQueryWizardContext } from "./PostgresQueryWizards/FunctionQueryWizard/IPostgresFunctionQueryWizardContext";
+import { runPostgresQueryWizard } from "./PostgresQueryWizards/runPostgresQueryWizard";
 
 export async function createPostgresFunctionQuery(context: IActionContext, treeItem?: PostgresFunctionsTreeItem): Promise<void> {
     const wizardContext: IPostgresFunctionQueryWizardContext = context;
@@ -22,11 +19,5 @@ export async function createPostgresFunctionQuery(context: IActionContext, treeI
         title: 'Create PostgreSQL Function Query'
     });
 
-    await wizard.prompt();
-    await wizard.execute();
-    await vscodeUtil.showNewFile(nonNullProp(wizardContext, 'query'), postgresBaseFileName, postgresFileExtension);
-
-    if (treeItem) {
-        await connectPostgresDatabase(wizardContext, treeItem.parent);
-    }
+    await runPostgresQueryWizard(wizard, wizardContext, treeItem);
 }
